Add PotentialEdgeMatrix class and .py file for the gluing classes
Adds a new class called PotentialEdgeMatrix to represent the matrix of variables.
This class is kinda bare bones at the moment so additional methods should be added as needed.
We should also eventually change the data structure used to represent the matrix, since as @davidtj2 pointed out, python lists are implemented as linked lists, which are not efficient at retrieving values.

Finally, also adds a `GluingClasses.py` and a `__init__.py` file so that the classes can be imported in other files.
We just need to be careful, since changing the jupyter notebook file for GluingClasses will not change how the code behaves in other files that import the classes.