Fix edge case of independent set for Km,Jn and add some additional comments
Fixes an issue where the vertex extension for Km,Jn was missing some of the possible neighborhoods when an independent set already had a vertex in the bottom of an interval.
Splits the handling of independent sets into two cases for this issue.
Also adds some (hopefully) more helpful comments to the code.